+++ Util/Docker/run.sh
@@ -20,7 +20,7 @@ User and group options:
 
 Ubuntu distribution:
 
-    --ubuntu-distro DISTRO   Specify ubuntu distro (default: 22.04)
+    --ubuntu-distro DISTRO   Specify ubuntu distro (default: 24.04, also supports 22.04)
 
 Build options:
 
@@ -32,7 +32,7 @@ Other commands:
 END
 )
 
-UBUNTU_DISTRO=22.04
+UBUNTU_DISTRO=24.04
 
 RUN_DEV=false